backend/pyproject.toml
[tool.poetry]
name = "AgenticShop-Backend"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = "^0.115.6"
llama-index-core = "^0.12.11"
alembic = "^1.14.1"
sqlalchemy = "^2.0.37"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pgvector = "^0.3.6"
pydantic-settings = "^2.8.1"
llama-index-embeddings-azure-openai = "^0.3.0"
llama-index-vector-stores-postgres = "^0.4.2"
llama-index-storage-chat-store-postgres = "^0.2.0"
llama-index-memory-mem0 = "^0.3.0"
sse-starlette = "^2.2.1"
llama-index = "^0.12.33"
arize-phoenix-otel = "^0.9.2"
openinference-instrumentation-llama-index = "^4.2.1"
arize-phoenix = "^8.31.0"
json5 = "^0.12.0"


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.3"
black = "^24.10.0"
flake8 = "^7.1.1"
pre-commit = "^4.0.1"
isort = "^5.13.2"
add-trailing-comma = "^3.1.0"
detect-secrets = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
